refactor(ndt7): avoid counterflow panic on closed channel
This diff defines a protocol for shutting down the sender where
we close the conn to signal goroutines we're done, then we
wait for them to join using a waitgroup, and only at this point
we close the channel.

Additionally, reckon that the counterflow reader may block
writing on a channel and change the write pattern to be
nonblocking and introduce a generous buffer to allow for
writing without losing data in normal conditions.

While there:

1. introduce a .gitignore file

2. propose an even clearer concurrency pattern

3. fix the build for macOS

4. introduce smoke testing at github